+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+# 0.2.3
+
+- Change folder to `buildSrc/src/main/kotlin` see [separate language source files](https://docs.gradle.org/current/userguide/organizing_gradle_projects.html#sec:separate_language_source_files)  
+- Create empty `buildSrc/settings.gradle.kts` if none exists
+- Use tasks.register to optimize configuration time
+
 # 0.2.2
 
 - `Versions` is now in a separate file `buildSrc/src/main/java/Versions.kt` (this way it works out of the box from a groovy build.gradle file)